trig ok + csv

+++ lab2/src/main/java/org/testing/math/log/AbsLogFunc.java
@@ -3,7 +3,7 @@ package org.testing.math.log;
 public abstract class AbsLogFunc {
     BaseLog ln;
 
-    public abstract Double solveLog(Long x, Long base, Long acc);
+    public abstract Double solveLog(double x, Long base, Long acc);
     public abstract void setLn(BaseLog ln);
 
 }
